Cast reconstructed attention weights
Agent-Logs-Url: https://github.com/gokamoda/transformer-feature-extractor/sessions/4ad929a6-201e-42b2-94db-07b8f2f860e9

+++ src/feature_extractor/reconstruction/attention.py
@@ -78,10 +78,8 @@ def reconstruct_attention_weights(
     attn_scores = _apply_causal_mask(attn_scores, mask_value)
 
     if architecture.attn_position_embeddings_arg_name is not None:
-        attn_weights = torch.softmax(attn_scores, dim=-1, dtype=torch.float32).to(
-            query.dtype
-        )
+        attn_weights = torch.softmax(attn_scores, dim=-1, dtype=torch.float32)
     else:
         attn_weights = torch.softmax(attn_scores, dim=-1)
 
-    return attn_weights
+    return attn_weights.to(query.dtype)
